fix(pix): make the CI clippy gate pass again
`cargo clippy --workspace --all-targets --features testing -- -D warnings`
(justfile:81, and what CI runs) failed on five pre-existing issues. Plain
`cargo clippy -F testing` misses all of them because they live in test and
bench targets.

The bench was outright broken: `PIX_THRESHOLD` was declared `u16` while
`DEFAULT_POLY_THRESHOLD` and `SsaGeneratorConfig::threshold` are both `u8`, so
`protocol_throughput_emulated_bench` had not compiled since the threshold was
re-tuned.

The three `vsss-rs` deprecations are pure renames — the deprecated
`split_secret_with_participant_generator` delegates to the plural form with an
identical signature, and `ParticipantIdGeneratorType` is a type alias for
`ParticipantIdGenerator` — so the call sites move with no behaviour change.

The remaining two are lints firing on deliberate code, and are allowed rather
than rewritten: `MinimalDeposit` *is* `()`, which is what makes passing it a
unit argument, and the buffer-size guard is constant by construction. Clippy's
`const` block alternative for the latter cannot format the offending value into
the message, which is the more useful half of the assertion.

+++ transport/hopr/benches/protocol_throughput_emulated_bench.rs
@@ -35,7 +35,7 @@ const SAMPLE_SIZE: usize = 50;
 ///
 /// This is what sets the per-share cost: `next_share` is a Horner evaluation over `threshold`
 /// coefficients.
-const PIX_THRESHOLD: u16 = DEFAULT_POLY_THRESHOLD;
+const PIX_THRESHOLD: u8 = DEFAULT_POLY_THRESHOLD;
 
 /// Whether the PIX share generator handed to the encoder has a committed SSA.
 ///
